backend/pyproject.toml
[project]
name = "image-hosting-app"
version = "0.1.0"
description = "Image Hosting API - Learning Distributed Systems"
# readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database
    "sqlalchemy>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Storage
    "aiofiles>=24.1.0",
    "minio>=7.2.0", # Phase 2: S3-compatible object storage
    # Caching (Phase 2)
    "redis>=5.2.0", # Redis client with async support
    "hiredis>=3.0.0", # C parser for performance
    # Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    # Utilities
    "python-dotenv>=1.0.1",
    "httpx>=0.28.0",
    "pillow>=12.0.0",
    "python-jose>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.3.0",
    # Phase 3: Web UI
    "jinja2>=3.1.0",
    "itsdangerous>=2.2.0", # For CSRF tokens
    "supabase>=2.27.1",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "aiosqlite>=0.20.0",
    "pillow>=11.0.0",         # For generating test images + Phase 1.5 dimensions
    "greenlet>=3.3.0",        # Required for SQLAlchemy async

    # Database (dev only - for Alembic migrations)
    "psycopg2-binary>=2.9.0", # Sync driver for Alembic

    # Code Quality
    "ruff>=0.8.0",
    "black>=24.10.0",
    "mypy>=1.13.0",

    # Type Stubs
    "types-aiofiles>=24.1.0",
]

# Optional dependencies for future phases
future = [
    "python-magic>=0.4.27",   # Phase 2: Advanced validation (deferred)
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by black)
    "B008",   # function call in default argument (FastAPI Depends pattern)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "detect-secrets>=1.5.0",
    "pre-commit>=4.5.1",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
